fix blessing metadata/MetaMap top level values
This branch is derived from and is compatible with
"updated Link/Image attrs support with backwards compatible JSON read/write"
https://github.com/bpj/Pandoc-Elements/commit/636b88fd778b49568f102ebaf759992c0d2b0a85

This commit fixes a bug I accidentally ran into when writing a
routine for mining metadata. Element hash values in the top-level
metadata hash or in a MetaMap wouldn't get blessed into their
proper packages, because Pandoc::Walker::transform() doesn't run
the action(s) on the top level of the data passed as its first
argument argument, but only on its children, and thus
Pandoc::Elements::pandoc_json() failed to bless the values of a
non-element hash, like the metadata hash itself or the values of
the content hash of MetaMap elements when transform called itself
recursively on the values of non-element hashes.

This commit modifies pandoc_json() so that it loops through the
top level metadata hash and any MetaMap objects and fixes this
specifically I'm thinking that maybe Pandoc::Walker::transform()
should look if its argument data is an element and run its action
on it in that case, but what should it then do with any return
value, especially if the return value is an array?

The included test meta-blessing.t fails on current master
https://github.com/nichtich/Pandoc-Elements/commit/fc41f52ae4ae76d16173b2daaf349adae0b6ebd6
but passes here.