specific TO_JSON methods to ensure right scalar types
Unlike Perl pandoc cares about whether a given scalar is a string
(including "quoted digits"), an integer or a float. In most
places pandoc expects strings, but there are some places where it
expects integers or floats. This commit modifies/adds TO_JSON
methods to ensure that scalars in all elements are of the right
type. It also ensures that the value of a MetaBool is a
JSON::PP::Boolean object, so that it gets serialized as a JSON Boolean.
It does so without modifying the data in the object to be
serialized. If anything in the object *is* modified that is a
bug. The reason for this is that a user may legitimately have an
object with an appropriate TO_JSON method take the place of a
scalar without that object being flattened just because they call
the (lowercase) to_json method. Also added: an xt test to ensure
that pandoc doesn't choke on the reserialized version of a
document with 'bad' scalars: all should-be numbers replaced by
strings, and all (one) digits-only strings replaced by a number.