Fix list index regex and _bad_pointer return value.
Metadata.pm: fix buggy list index match regex and
make sure _bad_pointer doesn't return an empty list.

The old regex for matching valid list indices -- `/^[1-9]*[0-9]$/ -- was buggy:
it rejected any indices containing non-final 0's, so it began to give false
negatives for 100-109 and similarly higher up. While it is perhaps unlikely
that there will be any lists that long in metadata those indices *are*
valid and should be accepted. The new regex is `/^[1-9][0-9]*$|^0$/`.
While it has lone zero in a separate branch it is reasonably fast because
the first branch will fail immediately on an initial 0 and the second
branch then tried immediately, failing if anything follows the 0.

In the course of fixing the above bug I noticed that _bad_pointer
returned nothing rather than an explicit undefined value in non-strict
mode, which led to wrong results when calling `value()` in list context,
since the return values for rejected pointers would just be missing.

I added tests for both thnese things to pointer.t